@@ -1,47 +1,31 @@
 package de.arthurpicht.barnacle.mapping;
 
 import de.arthurpicht.barnacle.configuration.GeneratorConfiguration;
-import de.arthurpicht.barnacle.context.GeneratorContext;
 
 import java.util.*;
 
-
-@SuppressWarnings("rawtypes")
 public class Entity {
 
-	private GeneratorConfiguration generatorConfiguration;
-	
-//	private String vofClassName;
-	private Class vofClass;
+	private final GeneratorConfiguration generatorConfiguration;
+	private final Class<?> vofClass;
+	private final List<Attribute> attributes;
+	private final Map<String, List<Attribute>> uniqueConstraints;
+	private final Map<String, ForeignKeyWrapper> foreignKeyConstraints;
 	private String tableName;
-	private List<Attribute> attributes;
-	private Map<String, List<Attribute>> uniqueConstraints;
-	private Map<String, ForeignKeyWrapper> foreignKeyConstraints;
 	private boolean vobFactoryMethod = false;
-	
 	private boolean isAssociationTable = false;
 	private ForeignKeyWrapper associationForeignKeyA = null;
 	private ForeignKeyWrapper associationForeignKeyB = null;
-	
-//	public Entity(String vofClassName) {
-//		this.vofClassName = vofClassName;
-//		this.attributes = new ArrayList<Attribute>();		
-//	}
 
-	public Entity(Class vofClass) {
-	    this.generatorConfiguration = GeneratorContext.getInstance().getGeneratorConfiguration();
-
+	public Entity(Class<?> vofClass, GeneratorConfiguration generatorConfiguration) {
+		this.generatorConfiguration = generatorConfiguration;
 		this.vofClass = vofClass;
-		this.attributes = new ArrayList<Attribute>();
-		this.uniqueConstraints = new HashMap<String, List<Attribute>>();
-		this.foreignKeyConstraints = new HashMap<String, ForeignKeyWrapper>();
+		this.attributes = new ArrayList<>();
+		this.uniqueConstraints = new HashMap<>();
+		this.foreignKeyConstraints = new HashMap<>();
 	}
 	
-//	public String getVofClassName() {
-//		return this.vofClassName;
-//	}
-	
-	public Class getVofClass() {
+	public Class<?> getVofClass() {
 		return this.vofClass;
 	}
 
@@ -50,7 +34,7 @@ public class Entity {
 	}
 	
 	public String getTableName() {
-		return tableName;
+		return this.tableName;
 	}
 	
 	public void addAttribute(Attribute attribute) {
@@ -58,17 +42,11 @@ public class Entity {
 	}
 
 	public List<Attribute> getAttributes() {
-		return attributes;
+		return this.attributes;
 	}
 	
-	/**
-	 * Returns all attributes except attributes annotated
-	 * as auto increment.
-	 * 
-	 * @return
-	 */
 	public List<Attribute> getNonAutoIncrementAttributes() {
-		List<Attribute> nonAutoIncrementAttributes = new ArrayList<Attribute>();
+		List<Attribute> nonAutoIncrementAttributes = new ArrayList<>();
 		for (Attribute attribute : this.attributes) {
 			if (!attribute.isAutoIncrement()) {
 				nonAutoIncrementAttributes.add(attribute);
@@ -77,12 +55,6 @@ public class Entity {
 		return nonAutoIncrementAttributes;
 	}
 	
-	/**
-	 * If existing returns the auto increment attribute.
-	 * Returns null in other cases.
-	 *  
-	 * @return
-	 */
 	public Attribute getAutoIncrementAttribute() {
 		for (Attribute attribute : this.attributes) {
 			if (attribute.isAutoIncrement()) {
@@ -92,12 +64,6 @@ public class Entity {
 		return null;
 	}
 	
-	/**
-	 * Determines the number of field defined as primary key
-	 * in this entity.
-	 * 
-	 * @return
-	 */
 	public int getNrPkAttributes() {
 		int nrPkAttributes = 0;
 		for (Attribute attribute : this.attributes) {
@@ -108,22 +74,12 @@ public class Entity {
 		return nrPkAttributes;
 	}
 	
-	/**
-	 * Determines whether primary key is a composed primary key.
-	 * 
-	 * @return
-	 */
 	public boolean isComposedPk() {
 		return this.getNrPkAttributes() > 1;
 	}
 	
-	/**
-	 * Returns all attributes annotated as primary key.
-	 * 
-	 * @return
-	 */
 	public List<Attribute> getPkAttributes() {
-		List<Attribute> pkAttributes = new ArrayList<Attribute>();
+		List<Attribute> pkAttributes = new ArrayList<>();
 		for (Attribute attribute : this.attributes) {
 			if (attribute.isPrimaryKey()) {
 				pkAttributes.add(attribute);
@@ -132,13 +88,8 @@ public class Entity {
 		return pkAttributes;
 	}
 	
-	/**
-	 * Returns all attributes NOT annotated as primary key.
-	 * 
-	 * @return
-	 */
 	public List<Attribute> getNonPkAttributes() {
-		List<Attribute> nonPkAttributes = new ArrayList<Attribute>();
+		List<Attribute> nonPkAttributes = new ArrayList<>();
 		for (Attribute attribute : this.attributes) {
 			if (!attribute.isPrimaryKey()) {
 				nonPkAttributes.add(attribute);
@@ -147,13 +98,6 @@ public class Entity {
 		return nonPkAttributes;
 	}
 	
-	/**
-	 * Returns attribute object by its field name. Returns null if no
-	 * attribute exists with this field name.
-	 * 
-	 * @param fieldName
-	 * @return
-	 */
 	public Attribute getAttributeByFieldName(String fieldName) {
 		for (Attribute attribute : this.attributes) {
 			if (attribute.getFieldName().equals(fieldName)) {
@@ -163,13 +107,6 @@ public class Entity {
 		return null;
 	}
 	
-	/**
-	 * Returns attribute object by its column name. Returns null if no
-	 * attribute exists with this column name.
-	 * 
-	 * @param columnName
-	 * @return
-	 */
 	public Attribute getAttributeByColumnName(String columnName) {
 		for (Attribute attribute : this.attributes) {
 			if (attribute.getColumnName().equals(columnName)) {
@@ -179,18 +116,11 @@ public class Entity {
 		return null;
 	}
 	
-	/**
-	 * Adds an attribute to the referenced unique index. If no index with
-	 * the given name is preexisting, a new index is created.
-	 * 
-	 * @param indexName
-	 * @param attribute
-	 */
 	public void addUniqueField(String indexName, Attribute attribute) {
 		
 		List<Attribute> uniqueAttributeList = this.uniqueConstraints.get(indexName);
 		if (uniqueAttributeList == null) {
-			uniqueAttributeList = new ArrayList<Attribute>();
+			uniqueAttributeList = new ArrayList<>();
 			uniqueAttributeList.add(attribute);
 			this.uniqueConstraints.put(indexName, uniqueAttributeList);
 		} else {
@@ -198,38 +128,14 @@ public class Entity {
 		}
 	}
 	
-	/**
-	 * Returns all names of the unique indices.
-	 * 
-	 * @return
-	 */
 	public Set<String> getAllUniqueIndicesNames() {
 		return this.uniqueConstraints.keySet();
 	}
 	
-	/**
-	 * Returns a foreign key by passed name. If requested foreign
-	 * key object does not exist, null is returned.
-	 * 
-	 * @param foreignKeyName
-	 * @return
-	 */
 	public ForeignKeyWrapper getForeignKeyByName(String foreignKeyName) {
-		ForeignKeyWrapper foreignKey = this.foreignKeyConstraints.get(foreignKeyName);
-		return foreignKey;
+		return this.foreignKeyConstraints.get(foreignKeyName);
 	}
-	
-	
-	/**
-	 * Returns a foreign key by passed name. If requested foreign
-	 * key object does not exist, a new one is created, stored and
-	 * given back. If the passed parameter foreignKeyName is empty,
-	 * a generated key name is used to create a new foreign key 
-	 * object.
-	 * 
-	 * @param foreignKeyName
-	 * @return
-	 */
+
 	public ForeignKeyWrapper getOrCreateForeignKeyByName(String foreignKeyName) {
 		ForeignKeyWrapper foreignKey = this.foreignKeyConstraints.get(foreignKeyName);
 		if (foreignKey == null) {
@@ -242,15 +148,9 @@ public class Entity {
 		return foreignKey;
 	}
 	
-	/**
-	 * Determines the next unused foreign key name following the
-	 * pattern 'fk_{tablename}_{1, 2, ...}' 
-	 * 
-	 * @return
-	 */
 	private String getNextForeignKeyName() {
 		int i = 0;
-		String nextForeignKeyName = new String();
+		String nextForeignKeyName;
 		ForeignKeyWrapper foreignKey;
 		do {
 			i++;
@@ -260,22 +160,12 @@ public class Entity {
 		return nextForeignKeyName;
 	}
 	
-	/**
-	 * Returns all names of defined foreign keys.
-	 * 
-	 * @return
-	 */
 	public Set<String> getAllForeignKeyNames() {
 		return this.foreignKeyConstraints.keySet();
 	}
 	
-	/**
-	 * Returns all foreign keys.
-	 * 
-	 * @return
-	 */
 	public Set<ForeignKeyWrapper> getAllForeignKeys() {
-		Set<ForeignKeyWrapper> foreignKeySet = new HashSet<ForeignKeyWrapper>();
+		Set<ForeignKeyWrapper> foreignKeySet = new HashSet<>();
 		for (String foreignKeyName : this.foreignKeyConstraints.keySet()) {
 			ForeignKeyWrapper foreignKeyWrapper = this.foreignKeyConstraints.get(foreignKeyName);
 			foreignKeySet.add(foreignKeyWrapper);
@@ -283,181 +173,91 @@ public class Entity {
 		return foreignKeySet;
 	}
 	
-	/**
-	 * Returns all foreign keys that reference this entity.
-	 * 
-	 * @return
-	 */
-	public Set<ForeignKeyWrapper> getAllReferencingForeignKeys() {
-		Set<ForeignKeyWrapper> referencingForeignKeyWrapper = new HashSet<ForeignKeyWrapper>();
-		for (Entity entity : EntityCollection.getEntities()) {
-			for (ForeignKeyWrapper foreignKeyWrapper : entity.getAllForeignKeys()) {
-				if (foreignKeyWrapper.getTargetEntity().getTableName().equals(this.getTableName())) {
-					referencingForeignKeyWrapper.add(foreignKeyWrapper);
-				}
-			}
-		}
-		return referencingForeignKeyWrapper;
-	}
+//	public Set<ForeignKeyWrapper> getAllReferencingForeignKeys() {
+//		Set<ForeignKeyWrapper> referencingForeignKeyWrapper = new HashSet<>();
+//		for (Entity entity : EntityCollection.getEntities()) {
+//			for (ForeignKeyWrapper foreignKeyWrapper : entity.getAllForeignKeys()) {
+//				if (foreignKeyWrapper.getTargetEntity().getTableName().equals(this.getTableName())) {
+//					referencingForeignKeyWrapper.add(foreignKeyWrapper);
+//				}
+//			}
+//		}
+//		return referencingForeignKeyWrapper;
+//	}
 	
-	/**
-	 * Returns all attributes assigned to given index.
-	 * 
-	 * @param indexName
-	 * @return
-	 */
 	public List<Attribute> getAttributesByUniqueIndexName(String indexName) {
 		return this.uniqueConstraints.get(indexName);
 	}
 	
-	/**
-	 * Checks whether value object contains a vob factory method. 
-	 * 
-	 * @return
-	 */
 	public boolean isVobFactoryMethod() {
 		return this.vobFactoryMethod;
 	}
 	
-	/**
-	 * Defines wheather value object contains a vob factory method.
-	 * 
-	 * @param vobFactoryMethod
-	 */
 	public void setVobFactoryMethod(boolean vobFactoryMethod) {
 		this.vobFactoryMethod = vobFactoryMethod;
 	}
 	
-	/**
-	 * Returns the simple name of the corresponding value object class.
-	 * 
-	 * @return
-	 */
 	public String getVoSimpleClassName() {
 		String vofSimpleClassName = this.vofClass.getSimpleName();
-		String voSimpleClassName = vofSimpleClassName.substring(0, vofSimpleClassName.length()-3) + "VO";
-		return voSimpleClassName;
+		return vofSimpleClassName.substring(0, vofSimpleClassName.length() - 3) + "VO";
 	}
 	
-	/**
-	 * Returns the canonical name of the corresponding value object class.
-	 * 
-	 * @return
-	 */
 	public String getVoCanonicalClassName() {
 		return this.generatorConfiguration.getVoPackageName() + "." + this.getVoSimpleClassName();
 	}
 	
-	/**
-	 * Returns the simple name of the corresponding value object business class.
-	 * 
-	 * @return
-	 */
 	public String getVobSimpleClassName() {
 		String vofSimpleClassName = this.vofClass.getSimpleName();
-		String vobSimpleClassName = vofSimpleClassName.substring(0, vofSimpleClassName.length()-3) + "VOB";
-		return vobSimpleClassName;
+		return vofSimpleClassName.substring(0, vofSimpleClassName.length() - 3) + "VOB";
 	}
 	
-	/**
-	 * Returns the canonical name of the corresponding value object business class.
-	 * 
-	 * @return
-	 */
 	public String getVobCanonicalClassName() {
 		return this.generatorConfiguration.getVobPackageName() + "." + this.getVobSimpleClassName();
 	}
 	
-	/**
-	 * Returns the simple name of the corresponding data access object class.
-	 * 
-	 * @return
-	 */
 	public String getDaoSimpleClassName() {
 		String vofSimpleClassName = this.vofClass.getSimpleName();
-		String daoSimpleClassName = vofSimpleClassName.substring(0, vofSimpleClassName.length()-3) + "DAO";
-		return daoSimpleClassName;
+		return vofSimpleClassName.substring(0, vofSimpleClassName.length()-3) + "DAO";
 	}
 	
-	/**
-	 * Returns the canonical name of the corresponding data access object class.
-	 * 
-	 * @return
-	 */
 	public String getDaoCanonicalClassName() {
 		return this.generatorConfiguration.getDaoPackageName() + "." + this.getDaoSimpleClassName();
 	}
 	
-	/**
-	 * Returns the simple name of the corresponding value field object class.
-	 * @return
-	 */
 	public String getVofSimpleClassName() {
 		return this.vofClass.getSimpleName();
 	}
 	
-	/**
-	 * Returns the simple name of the corresponding value object class 
-	 * containing the private key fields.
-	 * 
-	 * @return
-	 */
 	public String getPkSimpleClassName() {
 		String vofSimpleClassName = this.vofClass.getSimpleName();
-		String pkSimpleClassName = vofSimpleClassName.substring(0, vofSimpleClassName.length()-3) + "PK";
-		return pkSimpleClassName;
+		return vofSimpleClassName.substring(0, vofSimpleClassName.length()-3) + "PK";
 	}
 	
-	/**
-	 * Returns the canonical name of the corresponding value object class
-	 * containing the private key fields.
-	 * 
-	 * @return
-	 */
 	public String getPkCanonicalClassName() {
 		return this.generatorConfiguration.getVoPackageName() + "." + this.getPkSimpleClassName();
 	}
 
-	/**
-	 * Checks whether this entity represents an association-table.
-	 * 
-	 * @return
-	 */
 	public boolean isAssociationTable() {
 		return isAssociationTable;
 	}
 	
-	/**
-	 * Defines this entity as an representation for an association-table. 
-	 * 
-	 * @param isAssociationTable
-	 */
 	public void setAssociationTable(boolean isAssociationTable) {
 		this.isAssociationTable = isAssociationTable;
 	}
 	
-	/**
-	 * Returns foreignKey as first part of association. 
-	 * 
-	 * @return
-	 */
 	public ForeignKeyWrapper getAssociationForeignKeyA() {
-		return associationForeignKeyA;
+		return this.associationForeignKeyA;
 	}
 
 	/**
 	 * Sets foreignKey as first part of association.
-	 * 
-	 * @param associationForeignKeyA
 	 */
 	public void setAssociationForeignKeyA(ForeignKeyWrapper associationForeignKeyA) {
 		this.associationForeignKeyA = associationForeignKeyA;
 	}
 
 	/**
-	 * Returns foreignKey as seconf part of association.
-	 * 
-	 * @return
+	 * Returns foreignKey as second part of association.
 	 */
 	public ForeignKeyWrapper getAssociationForeignKeyB() {
 		return associationForeignKeyB;
@@ -465,44 +265,42 @@ public class Entity {
 
 	/**
 	 * Sets foreignKey as second part of association.
-	 * 
-	 * @param associationForeignKeyB
 	 */
 	public void setAssociationForeignKeyB(ForeignKeyWrapper associationForeignKeyB) {
 		this.associationForeignKeyB = associationForeignKeyB;
 	}
 
 	public String toString() {
-		String string = "Entity vobClassName=" + this.vofClass.getSimpleName() + " tableName=" + tableName;
-		string += "\n\tAttributes:";
+		StringBuilder string = new StringBuilder("Entity vobClassName=" + this.vofClass.getSimpleName() + " tableName=" + tableName);
+		string.append("\n\tAttributes:");
 		List<Attribute> attributeList = this.getAttributes();
 		for (Attribute attribute : attributeList) {
-			string += "\n\t\t" + attribute.toString();
+			string.append("\n\t\t").append(attribute.toString());
 		}
 		
-		string += "\n\tUnique Constraints:";
+		string.append("\n\tUnique Constraints:");
 		Set<String> uniqueKeyNames = this.uniqueConstraints.keySet();
 		for (String keyName : uniqueKeyNames) {
 			List<Attribute> uniqueAttributes = this.uniqueConstraints.get(keyName);
-			string += "\n\t\tkey name: " + keyName;
+			string.append("\n\t\tkey name: ").append(keyName);
 			for (Attribute attribute : uniqueAttributes) {
-				string += "\n\t\t" + attribute.toString();
+				string.append("\n\t\t").append(attribute.toString());
 			}
 		}
 		
 		Set<String> foreignKeyNames = this.foreignKeyConstraints.keySet();
 		for (String keyName : foreignKeyNames) {
 			ForeignKeyWrapper foreignKeyWrapper = this.foreignKeyConstraints.get(keyName);
-			string += "\n\t" + foreignKeyWrapper.toString();
+			string.append("\n\t").append(foreignKeyWrapper.toString());
 		}
 		
-		string += "\n\tIs Association Table?:" + this.isAssociationTable;
+		string.append("\n\tIs Association Table?:").append(this.isAssociationTable);
 		if (this.isAssociationTable) {
-			string += "\n\t\tForeignKeyA: " + this.associationForeignKeyA.toString();
-			string += "\n\t\tForeignKeyB: " + this.associationForeignKeyB.toString();
+			string.append("\n\t\tForeignKeyA: ").append(this.associationForeignKeyA.toString());
+			string.append("\n\t\tForeignKeyB: ").append(this.associationForeignKeyB.toString());
 		}
 		
-		return string;			
+		return string.toString();
 	}	
 
 }
