@@ -7,27 +7,27 @@ import java.util.Set;
  * Repräsentiert die Menge aller Entitäten.
  * Implementiert als Singleton.
  * 
- * @author Arthur Picht, Arthur Picht GmbH, (c) 2007
+ * @author Arthur Picht, (c) 2007 - 2023
  *
  */
 public class EntityCollection {
 
-	private static Set<Entity> entitySet;
-	
-	static {
-		entitySet = new HashSet<Entity>();
+	private final Set<Entity> entitySet;
+
+	public EntityCollection() {
+		this.entitySet = new HashSet<>();
+	}
+
+	public void addEntity(Entity entity) {
+		this.entitySet.add(entity);
 	}
 	
-	public static void addEntity(Entity entity) {
-		entitySet.add(entity);
+	public Set<Entity> getEntities() {
+		return this.entitySet;
 	}
 	
-	public static Set<Entity> getEntities() {
-		return entitySet;
-	}
-	
-	public static Entity getEntityByTableName(String tableName) {
-		for (Entity entity : entitySet) {
+	public Entity getEntityByTableName(String tableName) {
+		for (Entity entity : this.entitySet) {
 			if (entity.getTableName().equals(tableName)) {
 				return entity;
 			}
@@ -35,12 +35,12 @@ public class EntityCollection {
 		return null;
 	}
 	
-	public static String debugOut() {
-		String string = "EntityCollection";
-		for (Entity entity : entitySet) {
-			string += "\n" + entity.toString();
+	public String debugOut() {
+		StringBuilder string = new StringBuilder("EntityCollection");
+		for (Entity entity : this.entitySet) {
+			string.append("\n").append(entity.toString());
 		}
 		
-		return string;
+		return string.toString();
 	}
 }
